Split checkValidPriceRating() into untils dir
