+++ src/containers/RestaurantListContainer/restaurantListContainerUtils.test.js
@@ -2,7 +2,6 @@ import { faStar, faStarHalf } from "@fortawesome/free-solid-svg-icons";
 
 import {
   checkValidImageURL,
-  checkValidPriceRating,
   createIconArray,
   formattedAddress,
 } from "./restaurantListContainerUtils";
@@ -18,17 +17,6 @@ describe("checkValidImageURL tests", () => {
     const { invalidImageURLLabel } = LabelsObject;
     const invalidURL = "https://s3-media3.fl.yelpcdn.com/bphoto/None/o.jpg";
     expect(checkValidImageURL(invalidURL)).toEqual(invalidImageURLLabel);
-  });
-});
-
-describe("checkValidPriceRating tests", () => {
-  test("Returns price rating if available", () => {
-    const validPrice = "££";
-    expect(checkValidPriceRating(validPrice)).toEqual(validPrice);
-  });
-
-  test("Returns Price rating not available", () => {
-    expect(checkValidPriceRating(null)).toEqual("No price rating available");
   });
 });
