+++ src/utils/checkValidPriceRating.js
@@ -1,0 +1,9 @@
+const checkValidPriceRating = (price) => {
+  if (price === null) {
+    return "No price rating available";
+  } else {
+    return price;
+  }
+};
+
+export { checkValidPriceRating };
